requirements.txt
Flask
requests

google-cloud
pybigquery

python-telegram-bot
beautifulsoup4

pytube
Pillow
mutagen
sqlalchemy
